Add timing tests for escaping JSON strings
On my machine, the escaped test case has a runtime of 74ms. The chosen range of up to 5000ms provides a wide range that precludes the pathological case while still being acheivable on underpowered hardware.

Tests should be converted to proper benchmarks in a later PR (see #70).

Also exposes internals to `Chiron.Tests`.